Update callable format
Add ability for different functions for packing, unpacking, and making.
This should allow for a lot more customizability and control over what
is happening during each stage.

I think there are probably a few bugs that I still need to fix, but the
idea should be pretty much done and ready.

I will use some of these updates to test the elementnone that I've been
working on.